# lib/api-zod/src/generated/types/comune.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Modulistica Comuni - API per onoranze funebri
 * OpenAPI spec version: 0.1.0
 */

export interface Comune {
  id: number;
  nome: string;
  /** @nullable */
  provincia?: string | null;
  /** @nullable */
  orariApertura?: string | null;
  /** @nullable */
  necroscopica?: string | null;
  /** @nullable */
  anagrafeStatoCivile?: string | null;
  /** @nullable */
  poliziaMoreuaria?: string | null;
  /** @nullable */
  cimiteri?: string | null;
  /** @nullable */
  necroforo?: string | null;
  /** @nullable */
  impresaIncaricataTumulazione?: string | null;
  /** @nullable */
  tumulazioneCapellaPrivata?: string | null;
  /** @nullable */
  lapide?: string | null;
  /** @nullable */
  affissioni?: string | null;
  /** @nullable */
  luoghiDispersione?: string | null;
  /** @nullable */
  listinoDiTumulazione?: string | null;
  /** @nullable */
  appuntamentoAttoMorte?: string | null;
  /** @nullable */
  moduli?: string | null;
  /** @nullable */
  noteCostiPagamento?: string | null;
  createdAt: string;
  updatedAt: string;
}