# lib/api-zod/src/generated/types/moduloInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Modulistica Comuni - API per onoranze funebri
 * OpenAPI spec version: 0.1.0
 */
import type { ModuloInputEntityType } from './moduloInputEntityType';

export interface ModuloInput {
  /** @minLength 1 */
  nome: string;
  descrizione?: string;
  entityType: ModuloInputEntityType;
  entityId?: number;
  url?: string;
  note?: string;
}