# lib/api-zod/src/generated/types/moduloUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Modulistica Comuni - API per onoranze funebri
 * OpenAPI spec version: 0.1.0
 */
import type { ModuloUpdateEntityType } from './moduloUpdateEntityType';

export interface ModuloUpdate {
  /** @minLength 1 */
  nome?: string;
  descrizione?: string;
  entityType?: ModuloUpdateEntityType;
  entityId?: number;
  url?: string;
  note?: string;
}