# lib/api-zod/src/generated/types/nazioneUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Modulistica Comuni - API per onoranze funebri
 * OpenAPI spec version: 0.1.0
 */

export interface NazioneUpdate {
  /** @minLength 1 */
  nome?: string;
  consolato?: string;
  ambasciata?: string;
  telefono?: string;
  email?: string;
  documentiRichiesti?: string;
  procedureRimpatrio?: string;
  costiOrienttativi?: string;
  note?: string;
}